Island: Rename plugin_utils -> plugin_archive_parser
