+++ monkey/monkey_island/cc/repositories/plugin_archive_parser.py
